requirements.txt
#tokenizers==0.20.4
numpy==1.26.4
sentence-transformers~=3.3
#scikit-learn==1.5.2
#scipy==1.14.1
#transformers==4.47.0
jupyter~=1.1
ipykernel~=6.29
python-dotenv~=1.0
pydantic~=2.9
openai~=1.54
docarray~=0.40.0
pypdf~=5.1
lark~=1.2
tavily-python~=0.5.0
wikipedia~=1.4
chromadb~=0.5.18
httpx~=0.28.1
openai~=1.57

langchain~=0.3.10
langchain_core~=0.3.15
langchain_openai~=0.2.11
langchain-anthropic~=0.3.3
langchain_community~=0.3.10
langchain-chroma~=0.1.4
langchainhub~=0.1.21
langgraph~=0.2.56

azure-common~=1.1
azure-core~=1.31
azure-identity~=1.16
azure-search-documents~=11.4
azure-ai-inference~=1.0.0b6

fastapi[standard]~=0.115.5